chore: Move protocol errors and models out of evm
Move BytesConvertible to src.models

--- don't change below this line ---
ENG-3786 Took 39 minutes
