+++ src/evm/protocol/dodo/state.rs
@@ -16,13 +16,12 @@ use tycho_core::dto::ProtocolStateDelta;
 use crate::{
     evm::{
         protocol::{
-            errors::TransitionError,
             events::{EVMLogMeta, LogIndex},
-            models::GetAmountOutResult,
             state::{ProtocolEvent, ProtocolSim},
         },
         simulation::{SimulationEngine, SimulationParameters},
     },
+    protocol::{errors::TransitionError, models::GetAmountOutResult},
     u256_num::u256_to_f64,
 };
 
@@ -146,8 +145,8 @@ where
         token_in: &crate::models::ERC20Token,
         _token_out: &crate::models::ERC20Token,
     ) -> Result<
-        crate::evm::protocol::models::GetAmountOutResult,
-        crate::evm::protocol::errors::TradeSimulationError,
+        crate::protocol::models::GetAmountOutResult,
+        crate::protocol::errors::TradeSimulationError,
     > {
         let calldata = if self.base_token == token_in.address {
             self.helper_abi
