+++ src/lib.rs
@@ -20,6 +20,7 @@ pub use num_traits;
 #[cfg(feature = "evm")]
 pub mod evm;
 pub mod models;
+mod protocol;
 pub mod safe_math;
 pub mod serde_helpers;
 #[cfg(feature = "evm")]
