+++ src/protocol/errors.rs
